added use without ctpservices
