+++ app/idleMotions/conf/idleMotions.ini
@@ -1,7 +1,7 @@
-robot            cer
-period           1.0
-rpc_port         /idleMotions/rpc
-script_name      /home/user1/robotology/cer/app/idleMotions/scripts/r1_idleMotions.sh
+robot                       cer
+period                      1.0
+rpc_port                    /idleMotions/rpc
+use_ctpservice_or_setpos    ctpservice
 
 [MOTIONS]
 0               stretch_right_arm
@@ -10,3 +10,6 @@ script_name      /home/user1/robotology/cer/app/idleMotions/scripts/r1_idleMotio
 3               stretch_back
 4               look_gripper
 5               look_watch
+
+[USE_CTPSERVICES]
+ctps_script_name      /home/user1/robotology/cer/app/idleMotions/scripts/r1_idleMotions.sh
