+++ app/idleMotions/conf/idleMotions_SIM.ini
@@ -1,7 +1,7 @@
-robot           SIM_CER_ROBOT
-period          1.0
-rpc_port        /idleMotions/rpc
-script_name     /GIT/robotology/cer/app/idleMotions/scripts/r1_idleMotions.sh
+robot                       SIM_CER_ROBOT
+period                      1.0
+rpc_port                    /idleMotions/rpc
+use_ctpservice_or_setpos    ctpservice
 
 [MOTIONS]
 0               stretch_right_arm
@@ -10,3 +10,6 @@ script_name     /GIT/robotology/cer/app/idleMotions/scripts/r1_idleMotions.sh
 3               stretch_back
 4               look_gripper
 5               look_watch
+
+[USE_CTPSERVICES]
+script_name     /GIT/robotology/cer/app/idleMotions/scripts/r1_idleMotions.sh
